CUI-5: Block implicit form submission in the Select widget
Move the fix from the Form to the Select. The Select wraps react-select
in a display: contents div with an onKeyDown that preventDefaults Enter,
so Enter never escapes to trigger the browser's implicit submission —
regardless of whether the inner input is readonly or a writable search
field, and regardless of which form container the Select is in.

The previous Form-level fix only blocked Enter on readonly inputs, so a
searchable Select (>8 options) still submitted the form. It also did
nothing when the Select was used inside a consumer's own <form>.
